refctor deploy command
